implementing business rules

+++ src/main/java/br/alura/ForumHub/service/TopicoService.java
@@ -5,6 +5,7 @@ import br.alura.ForumHub.dto.topico.DadosTopicoCadastro;
 import br.alura.ForumHub.dto.topico.DadosTopicoResponse;
 import br.alura.ForumHub.model.entities.Curso;
 import br.alura.ForumHub.model.entities.Topico;
+import br.alura.ForumHub.model.entities.Usuario;
 import br.alura.ForumHub.repository.TopicoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -12,7 +13,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
 
 @Service
 public class TopicoService {
@@ -23,10 +23,15 @@ public class TopicoService {
     @Autowired
     private CursoService cursoService;
 
+    @Autowired
+    private UsuarioService usuarioService;
+
     @Transactional
     public DadosTopicoResponse criarTopico(DadosTopicoCadastro cadastro, Curso curso) {
 
-        Topico topico = new Topico(cadastro, curso);
+        Usuario autor = usuarioService.usuarioAtual();
+
+        Topico topico = new Topico(cadastro, curso, autor);
 
         Topico newTopico = topicoRepository.save(topico);
 
@@ -57,14 +62,26 @@ public class TopicoService {
             return null;
         }
 
+        if (topico.getAutor() != usuarioService.usuarioAtual()) {
+            return null;
+        }
+
         topico.atualizar(atualizacao);
 
         return new DadosTopicoResponse(topico);
     }
 
     @Transactional
-    public void removerTopico(Long id) {
+    public String removerTopico(Long id) {
+
+        Topico topico = topicoRepository.findById(id).orElse(null);
+
+        if (topico != null && topico.getAutor() != usuarioService.usuarioAtual()) {
+            return "Não foi possivel deletar o topico";
+        }
+
         topicoRepository.deleteById(id);
+        return null;
     }
 
 }
